# perf(fslib): run removePromise in parallel