--- a/packages/yarnpkg-fslib/sources/FakeFS.ts
+++ b/packages/yarnpkg-fslib/sources/FakeFS.ts
@@ -273,8 +273,13 @@ export abstract class FakeFS<P extends Path> {
     }
 
     if (stat.isDirectory()) {
-      if (recursive)
-        await Promise.all((await this.readdirPromise(p)).map(entry => this.removePromise(this.pathUtils.resolve(p, entry))));
+      if (recursive) {
+        const entries = await this.readdirPromise(p);
+
+        await Promise.all(entries.map(entry => {
+          return this.removePromise(this.pathUtils.resolve(p, entry));
+        }));
+      }
 
       // 5 gives 1s worth of retries at worst
       let t = 0;