**What's the problem this PR addresses?**

When `removePromise` encounters a folder it deletes its content one item at a time instead of all at once.

Ref https://github.com/yarnpkg/berry/issues/2575

**How did you fix it?**

Change the for loop to a `map` and `Promise.all`

**Result**

Running `hyperfine -w 5 'yarn lint-staged'` on https://github.com/jest-community/eslint-plugin-jest (cc @SimenB)

```diff
Benchmark #1: yarn lint-staged
-  Time (mean ± σ):      1.454 s ±  0.024 s    [User: 749.1 ms, System: 170.3 ms]
+  Time (mean ± σ):      1.458 s ±  0.041 s    [User: 732.1 ms, System: 199.7 ms]
-  Range (min … max):    1.420 s …  1.493 s    10 runs
+  Range (min … max):    1.396 s …  1.513 s    10 runs
```

Close to no difference in perfs but removing in parallel makes sense

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.